beaucoup de changements, et je suis trop fatigue pour le decouper. Y avait tellement de bugs que les comptes plus. J'ai corrige des problemes sur les liens, sur les sessions, sur les routes et sur les droits. J'ai aussi compris d'ou vient le stackoverflow qu'Andrei rencontre de temps en temps. Les filtres etant assez mal fichus dans rails, ils generent assez vite une grosse pile d'appels. J'ai un peu temporise le probleme en attendant de trouver une veritable correction. La pile d'appels devrait être suffisamment reduite.

+++ app/helpers/pages_helper.rb
@@ -18,18 +18,21 @@ module PagesHelper
   # Second, with an other controller :
   #   <%= link_to_show(edit_request_path(@request))%>
   def link_to_show(ar)
+    return nil unless ar
     url = (ar.is_a?(String) ? ar : { :action => 'show', :id => ar })
     link_to StaticImage::view, url, { :class => 'nobackground' }
   end
 
   # same behaviour as link_to_show
   def link_to_edit(ar)
+    return nil unless ar
     url = (ar.is_a?(String) ? ar : { :action => 'edit', :id => ar })
     link_to StaticImage::edit, url, { :class => 'nobackground' }
   end
 
   # same behaviour as link_to_show
   def link_to_delete(ar)
+    return nil unless ar
     url = (ar.is_a?(String) ? ar : { :action => 'destroy', :id => ar })
     link_to StaticImage::delete,  url, :class => 'nobackground', :method => :delete,
         :confirm => _('Do you really want to destroy this object ?')
